src/airflow/requirements.txt
apache-airflow-providers-microsoft-fabric>=1.0.0
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
mlflow>=2.10.0
requests>=2.31.0
